Allow capture of the unprocessed (raw) MPTS from appropriate input devices.
While Myth wont do any processing, if the input device supports pid
filtering, then that will still happen.  To make sure the entire stream is
capture, make sure a recording is enabled for each 'program' available on
the transport.

To enable this feature, set "Record Full TS?" to Yes in a recording profile
(in mythtv-setup) and then select that profile when creating the recording
rule.

This is based on the work by Daniel Kristjansson, just expanded to cover DVB
and HDHomeRun devices.  It can trivially be expanded to any other
appropriate input device.